+++ packages/node/src/integrations/undici/index.ts
@@ -3,9 +3,11 @@ import type { EventProcessor, Integration } from '@sentry/types';
 import {
   dynamicRequire,
   dynamicSamplingContextToSentryBaggageHeader,
+  getSanitizedUrlString,
+  parseUrl,
   stringMatchesSomePattern,
-  stripUrlQueryAndFragment,
 } from '@sentry/utils';
+import { LRUMap } from 'lru_map';
 
 import type { NodeClient } from '../../client';
 import { NODE_VERSION } from '../../nodeVersion';
@@ -29,12 +31,21 @@ export interface UndiciOptions {
    * Function determining whether or not to create spans to track outgoing requests to the given URL.
    * By default, spans will be created for all outgoing requests.
    */
-  shouldCreateSpanForRequest: (url: string) => boolean;
+  shouldCreateSpanForRequest?: (url: string) => boolean;
 }
 
 // Please note that you cannot use `console.log` to debug the callbacks registered to the `diagnostics_channel` API.
 // To debug, you can use `writeFileSync` to write to a file:
 // https://nodejs.org/api/async_hooks.html#printing-in-asynchook-callbacks
+//
+// import { writeFileSync } from 'fs';
+// import { format } from 'util';
+//
+// function debug(...args: any): void {
+//   // Use a function like this one when debugging inside an AsyncHook callback
+//   // @ts-ignore any
+//   writeFileSync('log.out', `${format(...args)}\n`, { flag: 'a' });
+// }
 
 /**
  * Instruments outgoing HTTP requests made with the `undici` package via
@@ -57,10 +68,13 @@ export class Undici implements Integration {
 
   private readonly _options: UndiciOptions;
 
+  private readonly _createSpanUrlMap: LRUMap<string, boolean> = new LRUMap(100);
+  private readonly _headersUrlMap: LRUMap<string, boolean> = new LRUMap(100);
+
   public constructor(_options: Partial<UndiciOptions> = {}) {
     this._options = {
       breadcrumbs: _options.breadcrumbs === undefined ? true : _options.breadcrumbs,
-      shouldCreateSpanForRequest: _options.shouldCreateSpanForRequest || (() => true),
+      shouldCreateSpanForRequest: _options.shouldCreateSpanForRequest,
     };
   }
 
@@ -85,6 +99,21 @@ export class Undici implements Integration {
       return;
     }
 
+    const shouldCreateSpan = (url: string): boolean => {
+      if (this._options.shouldCreateSpanForRequest === undefined) {
+        return true;
+      }
+
+      const cachedDecision = this._createSpanUrlMap.get(url);
+      if (cachedDecision !== undefined) {
+        return cachedDecision;
+      }
+
+      const decision = this._options.shouldCreateSpanForRequest(url);
+      this._createSpanUrlMap.set(url, decision);
+      return decision;
+    };
+
     // https://github.com/nodejs/undici/blob/e6fc80f809d1217814c044f52ed40ef13f21e43c/docs/api/DiagnosticsChannel.md
     ds.subscribe(ChannelName.RequestCreate, message => {
       const hub = getCurrentHub();
@@ -94,8 +123,8 @@ export class Undici implements Integration {
 
       const { request } = message as RequestCreateMessage;
 
-      const url = new URL(request.path, request.origin);
-      const stringUrl = url.toString();
+      const stringUrl = request.origin ? request.origin.toString() + request.path : request.path;
+      const url = parseUrl(stringUrl);
 
       if (isSentryRequest(stringUrl) || request.__sentry__ !== undefined) {
         return;
@@ -108,32 +137,41 @@ export class Undici implements Integration {
 
       if (activeSpan && client) {
         const clientOptions = client.getOptions();
-        const shouldCreateSpan = this._options.shouldCreateSpanForRequest(stringUrl);
 
-        if (shouldCreateSpan) {
+        if (shouldCreateSpan(stringUrl)) {
           const method = request.method || 'GET';
           const data: Record<string, unknown> = {
             'http.method': method,
           };
-          const params = url.searchParams.toString();
-          if (params) {
-            data['http.query'] = `?${params}`;
+          if (url.search) {
+            data['http.query'] = url.search;
           }
           if (url.hash) {
             data['http.fragment'] = url.hash;
           }
           const span = activeSpan.startChild({
             op: 'http.client',
-            description: `${method} ${stripUrlQueryAndFragment(stringUrl)}`,
+            description: `${method} ${getSanitizedUrlString(url)}`,
             data,
           });
           request.__sentry__ = span;
 
-          const shouldPropagate = clientOptions.tracePropagationTargets
-            ? stringMatchesSomePattern(stringUrl, clientOptions.tracePropagationTargets)
-            : true;
+          const shouldAttachTraceData = (url: string): boolean => {
+            if (clientOptions.tracePropagationTargets === undefined) {
+              return true;
+            }
 
-          if (shouldPropagate) {
+            const cachedDecision = this._headersUrlMap.get(url);
+            if (cachedDecision !== undefined) {
+              return cachedDecision;
+            }
+
+            const decision = stringMatchesSomePattern(url, clientOptions.tracePropagationTargets);
+            this._headersUrlMap.set(url, decision);
+            return decision;
+          };
+
+          if (shouldAttachTraceData(stringUrl)) {
             request.addHeader('sentry-trace', span.toTraceparent());
             if (span.transaction) {
               const dynamicSamplingContext = span.transaction.getDynamicSamplingContext();
@@ -155,8 +193,7 @@ export class Undici implements Integration {
 
       const { request, response } = message as RequestEndMessage;
 
-      const url = new URL(request.path, request.origin);
-      const stringUrl = url.toString();
+      const stringUrl = request.origin ? request.origin.toString() + request.path : request.path;
 
       if (isSentryRequest(stringUrl)) {
         return;
@@ -196,8 +233,7 @@ export class Undici implements Integration {
 
       const { request } = message as RequestErrorMessage;
 
-      const url = new URL(request.path, request.origin);
-      const stringUrl = url.toString();
+      const stringUrl = request.origin ? request.origin.toString() + request.path : request.path;
 
       if (isSentryRequest(stringUrl)) {
         return;